move benchmarking utilities to 'common' folder to clean up folder
