+++ bench/common/cli_benchmark_interface.hpp
