+++ bench/common/cli_device_selector.hpp
